+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.3.0 (Next)
 - Support Vault Kubernetes authentication method.
+- Support Vault AppRole authentication method.
 - `auth_engine` source parameter now required and not optional.
 - Generalize authentication parameters to `auth_mount` and `vault_role`.